В модели баннеров тип поля price изменен на IntegerField для представления цены в рублях без копеек.

+++ market/market_app/models.py
@@ -4,8 +4,8 @@ from django.db import models
 class Banner(models.Model):
     """ Баннер """
     title = models.CharField(max_length=50, verbose_name='заголовок')
-    # Максимальная цена 999 999 999.99
-    price = models.DecimalField(max_digits=11, decimal_places=2, verbose_name='цена')
+    # Цены в рублях (без копеек)
+    price = models.IntegerField(verbose_name='цена')
     image = models.CharField(max_length=300, verbose_name='изображение')
     image_alt = models.CharField(max_length=100, verbose_name='подсказка')
     link = models.CharField(max_length=300, verbose_name='url')
